+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add `find_csi_camera_index()` and support for using CSI cameras on Raspberry Pi 5.
 - Add `depad` option to `LibcameraCapture` to enable or disable stripping per-line stride padding.
 
 ## 2.17.0 (2026-06-02)